Modify: urls.py, views.py, utils.py, created decorator.py
