+++ users/utils.py
@@ -1,11 +1,6 @@
-import re, jwt
-
-from functools import wraps
-
-from django.http import JsonResponse
+import re
 
 from users.models import User
-from my_settings  import SECRET_KEY, ALGORITHM
 
 def is_valid(data):
     email    = data["email"]
@@ -29,22 +24,3 @@ def is_valid(data):
     return False
 
 
-def login_decorator(func):
-    @wraps(func)
-    def wrapper(self, request, *args, **kwargs):
-        try:
-            access_token = request.headers.get('Authorization', None)
-            payload      = jwt.decode(access_token, SECRET_KEY, ALGORITHM)
-            user         = User.objects.get(id = payload["id"])
-            request.user = user
-            
-            return func(self, request, *args, **kwargs)
-
-        except jwt.DecodeError:
-            return JsonResponse({"message" : "INVALID_TOKEN"}, status = 401)
-        
-        except User.DoesNotExist:
-            return JsonResponse({"message" : "INVALID_USER"}, status = 401)
-
-    return wrapper
-
